[CHG] Remove dependency on project. Modify module structure. Imporve readme file.
