+++ web_timeline/__openerp__.py
@@ -1,16 +1,24 @@
+# -*- coding: utf-8 -*-
+# Copyright 2016 ACSONE SA/NV (<http://acsone.eu>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 {
     'name': "Web timeline",
     'summary': """
-    Interactive visualization chart to visualize events in time
+        Interactive visualization chart to visualize events in time
     """,
-    "version": "0.1",
-    "author": "ACSONE SA/NV",
-    "category": "Acsone",
+    "version": "8.0.1.0.0",
+    'author': 'ACSONE SA/NV,'
+              'Odoo Community Association (OCA)',
+    "category": "Tools",
     "website": "http://acsone.eu",
-    'depends': ['web', 'project'],
-    'qweb': ['static/src/xml/web_timeline.xml'],
+    'depends': [
+        'web'
+    ],
+    'qweb': [
+        'static/src/xml/web_timeline.xml',
+    ],
     'data': [
         'views/web_timeline.xml',
-        'project_view.xml',
     ],
 }
